Update K.adoc
@wmo-im/tt-wismd  Please see update for KPI K as discussed in the meeting. 

anchor's to implement WMOCodes can be defined in 4 places,  If I give it 4 points then all the weight of the KPI goes to defining anchors which is not really the spirit of this KPI. 
I suggest to leave 1 point.  
any comments ? 

ps: I also updated the iso codelist in the examples to the authoritative one recommended  as the previous one is now dead.